+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Renderer is not always created first time [#1578](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1578)
 - Handle non-property fields, such as `PK` [#1580](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1580)
 - Added `Navigate` to the Class for the model variable [#1583](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1583)
+- Display extension scope specific icons for model [#1584](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1584)
 
 ### `Type System` enhancements
 - Introduced custom icon for `dynamic` attributes [#1582](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1582)